requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
beautifulsoup4>=4.12.2
google-search-results>=2.4.2
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
aiofiles>=23.0.0